chor(android): Move native implementation to MediaService
