+++ demo/src/main-view-model.ts
@@ -220,7 +220,8 @@ export class HelloWorldModel extends Observable implements PlaybackEventListener
     // console.log(`Player position: ${this.player.getCurrentPlaylistIndex()}@${this.player.getCurrentTime()} / ${this.player.getDuration()}`);
   }
 
-  public onPlaybackEvent(evt: PlaybackEvent) {/*
+  public onPlaybackEvent(evt: PlaybackEvent) {
+    /*
     if (evt === PlaybackEvent.SleepTimerChanged) {
       console.log(`SleepTimerChanged: ${this.player.getSleepTimerRemaining()}`);
     } else {
