+++ src/audioplayer-common.ts
@@ -1,62 +1,6 @@
 import { isIOS } from 'tns-core-modules/platform';
 import * as trace from 'tns-core-modules/trace';
-
-export const notaAudioCategory = 'NotaAudioPlayer';
-
-export class MediaTrack {
-  constructor(url: string, title: string, artist: string, album: string, albumArtUrl: string) {
-    this.url = url;
-    this.title = title;
-    this.artist = artist;
-    this.album = album;
-    this.albumArtUrl = albumArtUrl;
-  }
-  public url: string;
-  public title: string;
-  public artist: string;
-  public album: string;
-  public albumArtUrl: string;
-}
-
-export class Playlist {
-  public tracks: MediaTrack[];
-  public get length(): number {
-    return this.tracks.length;
-  }
-  constructor(public UID: string, ...tracks: MediaTrack[]) {
-    this.UID = UID;
-    this.tracks = tracks;
-  }
-}
-
-export enum PlaybackEvent {
-  Stopped,
-  Buffering,
-  Playing,
-  Paused,
-  EndOfTrackReached,
-  EndOfPlaylistReached,
-  EncounteredError,
-  TimeChanged,
-  SleepTimerChanged,
-  WaitingForNetwork,
-}
-
-export interface Config {
-  /**
-   * Max number of retry attempts before considering streaming failed
-   */
-  maxNetworkRetryCount: number;
-
-  /**
-   * Required number of seconds buffered before starting playback
-   */
-  requiredPrebufferSizeInSeconds: number;
-}
-
-export interface PlaybackEventListener {
-  onPlaybackEvent(evt: PlaybackEvent, arg?: any): void;
-}
+import { notaAudioCategory, PlaybackEvent, PlaybackEventListener, Playlist } from './audioplayer.types';
 
 let instanceNo = 0;
 export abstract class CommonAudioPlayer {
@@ -69,7 +13,7 @@ export abstract class CommonAudioPlayer {
   public playlist: Playlist;
 
   protected _queuedSeekTo: number = null;
-  protected _listener: PlaybackEventListener;
+  private _listener: PlaybackEventListener;
   protected seekIntervalSeconds = 15;
 
   public abstract isReady: Promise<any>;
@@ -174,7 +118,7 @@ export abstract class CommonAudioPlayer {
     return this.playlist ? this.playlist.UID : null;
   }
 
-  protected _onPlaybackEvent(evt: PlaybackEvent, arg?: any) {
+  public _onPlaybackEvent(evt: PlaybackEvent, arg?: any) {
     if (this._listener) {
       this._listener.onPlaybackEvent(evt, arg);
     }
